Add docstring and add hot_run parameter when downloading files for test purposes

+++ scripts/retrieve_databundle_light.py
@@ -61,7 +61,25 @@ def load_databundle_config(path):
     
     return config
 
-def download_and_unzip(host, config, rootpath, dest_path):
+"""
+    download_and_unzip(host, config, rootpath, dest_path, hot_run=True)
+
+Function to download and unzip the data by category
+
+Inputs
+------
+host : str
+    Name of the hosting platform: zenodo or google
+config : Dict
+    Configuration data for the category to download
+rootpath : str
+    Absolute path of the repository
+hot_run : Bool (default True)
+    When true the data are downloaded
+    When false, the workflow is run without downloading and unzipping
+
+"""
+def download_and_unzip(host, config, rootpath, hot_run=True):
     """
     Function to download and unzip data depending on the hosting platform.
     Currently, hosts accepted: zenodo and google
@@ -71,13 +89,14 @@ def download_and_unzip(host, config, rootpath, dest_path):
 
     if host=="zenodo":
         url=config["urls"]["zenodo"]
-        progress_retrieve(url, file_path)
-        logger.info(f"Extracting resources")
-        with ZipFile(file_path, "r") as zipObj:
-            # Extract all the contents of zip file in current directory
-            zipObj.extractall(path=dest_path)
-        os.remove(file_path)
-        logger.info(f"Download resource '{resource}' from cloud '{url}'.")
+        if hot_run:
+            progress_retrieve(url, file_path)
+            logger.info(f"Extracting resources")
+            with ZipFile(file_path, "r") as zipObj:
+                # Extract all the contents of zip file in current directory
+                zipObj.extractall(path=config["destination"])
+            os.remove(file_path)
+            logger.info(f"Download resource '{resource}' from cloud '{url}'.")
         return True
     elif host=="google":
 
@@ -97,18 +116,19 @@ def download_and_unzip(host, config, rootpath, dest_path):
         # get file id
         file_id = code_split[-1]
 
-        if os.path.exists(file_path):
+        if hot_run:
+            if os.path.exists(file_path):
+                os.remove(file_path)
+            gdd.download_file_from_google_drive(
+                file_id=file_id,
+                dest_path=file_path,
+                showsize=True,
+                unzip=False,
+            )
+            with ZipFile(file_path, "r") as zipObj:
+                # Extract all the contents of zip file in current directory
+                zipObj.extractall(path=config["destination"])
             os.remove(file_path)
-        gdd.download_file_from_google_drive(
-            file_id=file_id,
-            dest_path=file_path,
-            showsize=True,
-            unzip=False,
-        )
-        with ZipFile(file_path, "r") as zipObj:
-            # Extract all the contents of zip file in current directory
-            zipObj.extractall(path=dest_path)
-        os.remove(file_path)
         logger.info(f"Download resource '{resource}' from cloud '{url}'.")
 
         return True
@@ -196,10 +216,9 @@ if __name__ == "__main__":
     # download the selected bundles
     for b_name in bundle_to_download:
         host_list = config_bundles[b_name]["urls"]
-        dest_path = os.path.abspath(config_bundles[b_name]["destination"])
         # loop all hosts until data is successfully downloaded
         for host in host_list:
-            if download_and_unzip(host, config_bundles[b_name], rootpath, dest_path):
+            if download_and_unzip(host, config_bundles[b_name], rootpath):
                 break
 
     logger.info("Bundle successfully loaded and unzipped:\n\t" + "\n\t".join(bundle_to_download))
